+++ web/src/features/interview-session/client/components/interview-summary-input.tsx
@@ -1,7 +1,9 @@
 "use client";
 
+import { useState } from "react";
 import type { PromptInputMessage } from "@/components/ai-elements/prompt-input";
 import { Button } from "@/components/ui/button";
+import { InterviewPublicConsentModal } from "@/features/interview-report/client/components/interview-public-consent-modal";
 import { useInterviewCompletion } from "../hooks/use-interview-completion";
 import { InterviewChatInput } from "./interview-chat-input";
 
@@ -10,7 +12,6 @@ interface InterviewSummaryInputProps {
   input: string;
   onInputChange: (value: string) => void;
   onSubmit: (message: PromptInputMessage) => void;
-  onComplete: (reportId: string | null) => void;
   isLoading: boolean;
   error: Error | null | undefined;
 }
@@ -20,21 +21,20 @@ export function InterviewSummaryInput({
   input,
   onInputChange,
   onSubmit,
-  onComplete,
   isLoading,
   error,
 }: InterviewSummaryInputProps) {
-  const { isCompleting, completeError, handleAgree } = useInterviewCompletion({
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const { isCompleting, completeError, handleSubmit } = useInterviewCompletion({
     sessionId,
-    onComplete,
   });
 
   return (
     <>
       {!isLoading && (
         <div className="mb-3 flex flex-col gap-2">
-          <Button onClick={handleAgree} disabled={isCompleting}>
-            {isCompleting ? "送信中..." : "レポート内容に同意する"}
+          <Button onClick={() => setIsModalOpen(true)} disabled={isCompleting}>
+            {isCompleting ? "送信中..." : "レポート内容に同意して提出"}
           </Button>
           {completeError && (
             <p className="text-sm text-red-500">{completeError}</p>
@@ -48,6 +48,12 @@ export function InterviewSummaryInput({
         placeholder="レポートの修正要望があれば入力してください"
         isResponding={isLoading}
         error={error}
+      />
+      <InterviewPublicConsentModal
+        open={isModalOpen}
+        onOpenChange={setIsModalOpen}
+        onSubmit={handleSubmit}
+        isSubmitting={isCompleting}
       />
     </>
   );
