+++ web/src/features/interview-session/client/utils/interview-api-client.ts
@@ -1,5 +1,6 @@
 interface CompleteInterviewParams {
   sessionId: string;
+  isPublic: boolean;
 }
 
 interface CompleteInterviewResult {
